Clear out search text on resume
When the app starts or resumes, the search text will always be cleared out.  This should speed up consecutive searches, by not requiring that the user clear it explicitly.

This has two side effects: (1) that device rotation the device will clear the search text, and (2) interrupting Minimalist Search with another app (say, a phone call) will clear the search text.  In the interests of generally higher speed, and of a consistent interface, these two side effects are not a priority for this app.